[AMQ-8354] Resyncronization step5: Extract ActiveMQTextMessage based logic to a separate base class.
