+++ activemq-broker/src/main/java/org/apache/activemq/replica/storage/ReplicaRoleStorage.java
@@ -22,24 +22,18 @@ import org.apache.activemq.command.ActiveMQTextMessage;
 import org.apache.activemq.replica.util.ReplicaInternalMessageProducer;
 import org.apache.activemq.replica.ReplicaReplicationDestinationSupplier;
 import org.apache.activemq.replica.util.ReplicaRole;
-import org.apache.activemq.util.IdGenerator;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.List;
 
-public class ReplicaRoleStorage extends ReplicaBaseStorage {
-
-    private final Logger logger = LoggerFactory.getLogger(ReplicaRoleStorage.class);
+public class ReplicaRoleStorage extends ReplicaBaseTextStorage {
 
     public ReplicaRoleStorage(Broker broker, ReplicaReplicationDestinationSupplier destinationSupplier,
             ReplicaInternalMessageProducer replicaInternalMessageProducer) {
         super(broker, replicaInternalMessageProducer, destinationSupplier.getRoleQueue(),  "ReplicationPlugin.ReplicaFailOverStorage", null);
-        this.replicationProducerId.setConnectionId(new IdGenerator().generateId());
     }
 
     public ReplicaRole initialize(ConnectionContext connectionContext) throws Exception {
-        List<ActiveMQTextMessage> allMessages = super.initializeBase(connectionContext, true);
+        List<ActiveMQTextMessage> allMessages = super.initialize(connectionContext, true);
 
         if (allMessages.isEmpty()) {
             return null;
